refactor: extract result handling
