+++ src/admin/ts/single/single-generation-utils.ts
@@ -1,20 +1,7 @@
-export const REST_ENDPOINT =
-  (window as any).nuclenAdminVars?.rest_receive_content ||
-  '/wp-json/nuclear-engagement/v1/receive-content';
-
-export const REST_NONCE = (window as any).nuclenAdminVars?.rest_nonce || '';
-
-export function alertApiError(errMsg: string): void {
-  if (errMsg.includes('Invalid API key')) {
-    alert('Your API key is invalid. Please go to the Setup page and enter a new one.');
-  } else if (errMsg.includes('Invalid WP App Password')) {
-    alert('Your WP App Password is invalid. Please go to the Setup page and re-generate it.');
-  } else if (errMsg.includes('Not enough credits')) {
-    alert('Not enough credits for single-post generation.');
-  } else {
-    alert('Error: ' + errMsg);
-  }
-}
+export {
+  nuclenAlertApiError as alertApiError,
+  nuclenStoreGenerationResults as storeGenerationResults,
+} from '../generation/results';
 
 export function populateQuizMetaBox(postResult: any, finalDate?: string): void {
   const { date, questions } = postResult;
@@ -82,17 +69,3 @@ export function populateSummaryMetaBox(postResult: any, finalDate?: string): voi
   }
 }
 
-export async function storeGenerationResults(workflow: string, results: any) {
-  const payload = { workflow, results };
-  const resp = await fetch(REST_ENDPOINT, {
-    method: 'POST',
-    headers: {
-      'Content-Type': 'application/json',
-      'X-WP-Nonce': REST_NONCE,
-    },
-    credentials: 'include',
-    body: JSON.stringify(payload),
-  });
-  const data = await resp.json();
-  return { ok: resp.ok, data };
-}
